toyota stations by radius filter added

+++ trunk/ToyotaPOC/platforms/android/assets/www/js/app/locator/locator.controllers.js
@@ -14,15 +14,26 @@ locatorControllers.controller('LocatorController', function($scope)
 			}
 		});	
 		    
+			// define list load function
+			$scope.loadFavs = function()
+			{	
+				var positions = getPositions();
+			   	if (positions == null)
+			    positions = [];
+			    	        
+			    $scope.serviceList = positions;
+			}
+		    
+			
 		   document.addEventListener("pageinit", function(){			   
 			   //get entries from DB
-			    loadFavs();
+			    $scope.loadFavs();
 		   });
 			   
-		    
+		      
 		   $scope.init = function()
 		   {
-			   loadFavs();
+			   $scope.loadFavs();
 		   }
 		   
 		   $scope.init();
@@ -33,14 +44,7 @@ locatorControllers.controller('LocatorController', function($scope)
 
 
 
-function loadFavs()
-{	
-	  var positions = getPositions();
-	  if (positions == null)
-      positions = [];
- 	        
-      $scope.serviceList = positions;
-}
+
       
       function getPositions()
       {
@@ -140,6 +144,24 @@ function loadFavs()
       }
       
       
+		$scope.removeServiceStation = function(address) {
+			
+		    var positions = getPositions();
+		    	 	  
+		    for(var i = 0; i < positions.length; i++)
+		    {
+		    	if(positions[i].address == address)
+		    		positions.splice(i,1);//remove element at position i
+		    }
+		    	 	  
+		    var _db = window.localStorage;
+		    _db.setItem('positions', JSON.stringify(positions));
+		   		    
+		    $scope.loadFavs();
+		}
+      
+      
+      
 });  // end of controller
 
 
@@ -226,22 +248,36 @@ locatorControllers.controller('LookupController', function($scope)
     	  var long = $scope.longitude;
     	  var address = $scope.address;
     	 	
-    	  alert('saving '+ latitude + long + address);
     	 	  var positions = getPositions();
     	 	  if (positions == null)
     	 		  positions = [];
     	 	  
-    	 	  positions.unshift(new ServicePosition(latitude,long, address, new Date()));
+    	 	  if(! $scope.favExists(address, positions))
+    	 	  {
+    	 	     positions.unshift(new ServicePosition(latitude,long, address, new Date()));
     	 	 
-    	 	  if(positions.length >= MAX_POSITIONS)
-    	 		  positions.slice();
+    	 	     if(positions.length >= MAX_POSITIONS)
+    	 		    positions.slice();
     	 	  
-    	 	  var _db = window.localStorage;
-    	 	  _db.setItem('positions', JSON.stringify(positions));
-    	 	  alert('saved' +  JSON.stringify(positions));    	 
+    	 	     var _db = window.localStorage;
+    	 	     _db.setItem('positions', JSON.stringify(positions));
+    	 	     alert('saved' +  JSON.stringify(positions)); 
+    	 	  }
+    	 	  else // this address is already in the favourites
+    	 	  {
+    	 		  alert('Address ' + address + ' has been saved previously');
+    	 	  }
      }
     
-     
+     $scope.favExists = function(address, positions)
+     {
+    	 for(var i = 0; i < positions.length; i++)
+    	 {
+    		 if(positions[i].address == address) return true;
+    	 }
+    	 
+    	 return false;
+     }
      
 
      function getPositions()
@@ -255,6 +291,105 @@ locatorControllers.controller('LookupController', function($scope)
      		  return positions;
       }
 
+     
+     
+      $scope.searchStations = function(radius)
+      {
+    	  navigator.geolocation.getCurrentPosition(
+    			  function(result)
+    			  {
+    				  var coords = result.coords;
+    				  var myLat = coords.latitude;
+    				  var myLon = coords.longitude;
+    			      printCurrentPosition(coords);// print current address and long lat
+    			      
+    			      var canvas = document.getElementById('lookupCanvas');	
+    			      // Set the initial Lat and Long of the Google Map
+    			      var myLatLng = new google.maps.LatLng(myLat, myLon);   
+    			        
+    			        
+    			      // Google Map options
+    			      var myOptions = 
+    			      {
+    			        zoom: 15,
+    			        center: myLatLng,
+    			        mapTypeId: google.maps.MapTypeId.ROADMAP
+    			      };
+
+    			       var mapBounds = new google.maps.LatLngBounds();// create bounds obj
+    			       var map = new google.maps.Map(canvas, myOptions);
+    			       var toyotaStations = $scope.getToyotaStationsWithinRadius(radius, myLat, myLon); 
+    			   
+    			    	for(var i = 0; i < toyotaStations.length; i++)
+    			    	{
+    			    		var marker = new google.maps.Marker({
+    			    		position: toyotaStations[i],
+    			    	    map: map,
+    			    	    title: 'Current position'});
+    			    		
+    			    		mapBounds.extend(toyotaStations[i]);
+    			    	}
+    			      map.fitBounds(mapBounds);
+    			  
+    			  }, 
+    			  
+    			  function()
+    			  {
+    				  alert('failed to read current position');
+    			  }, 
+    	
+    			  { timeout: 60000, maximumAge:60000, enableHighAccuracy: true }
+    			  
+    	  );}
+
+      
+      
+      $scope.getToyotaStationsWithinRadius = function(radius, myLat, myLon)
+      {
+    	  var stations = getPositions();// change to retrieve data from database
+    	  var selectedStations = [];
+    	  
+    	  for(var i = 0; i < stations.length; i++)
+    	  {
+    		  var destLat = stations[i].latitude;
+    	      var destLon = stations[i].long;
+    		 
+    	      var distance = getDistance(myLat, myLon, destLat, destLon);
+    		  if(distance <= radius)
+    		  {
+    			  selectedStations.push(new google.maps.LatLng(destLat, destLon));
+    		  }
+    	  }
+
+    	  return selectedStations;
+      }
+     
+     
+      
+      function getDistance(lat1, lon1, lat2, lon2)
+      {
+
+    	 var R = 6371; // km
+    	 var dLat = (lat2 - lat1) * (Math.PI / 180);
+    	 var dLon = (lon2 - lon1) * (Math.PI / 180);
+    	 var lat1 = lat1 * (Math.PI / 180);
+    	 var lat2 = lat2 * (Math.PI / 180);
+
+    	 var a = Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+    	              Math.sin(dLon / 2) * Math.sin(dLon / 2) * Math.cos(lat1) * Math.cos(lat2);
+    	 var c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+    	 var d = R * c;
+    	 
+    	 return d.toFixed(2);    	  
+      }
+      
+      
+      
+      
+      
+      
+      
+     
      function ServicePosition(latitude,long, address, datetime)
      {
      	  this.latitude = latitude;
